Definition of I2C bus-type per-board. Segregated busio tests into i2c vs uart
